+++ components/response-form-base/question-element/question-single-choice-input.vue
@@ -1,10 +1,10 @@
 <script lang="ts" setup>
-import type { ResponseSettingsSingleChoice } from '~/components/form-response-base/form-response-base-settings';
+import type { ResponseSettingsSingleChoice } from '~/components/response-form-base/response-form-base-settings';
 import type {
   ResponseFormQuestionInvalid,
   ResponseFormQuestionSettingsBase,
   ResponseFormQuestionSettingsSingleChoice,
-} from '~/components/form-response-base/questionnaire-settings';
+} from '~/components/response-form-base/questionnaire-settings';
 
 type ModelValue = ResponseSettingsSingleChoice &
   ResponseFormQuestionSettingsSingleChoice &
@@ -71,3 +71,4 @@ const optionIds = question.value.options.map(() => createId());
 }
 </style>
 ~/components/form-response-base/questionnaire-settings
+~/components/response-form-base/form-response-base-settings~/components/response-form-base/questionnaire-settings
