+++ pages/questionnaires/[questionnaireID]/responses/new.vue
@@ -1,6 +1,5 @@
 <script lang="ts" setup>
-import type { ResponseFormQuestionnaireFormSettings } from '~/components/form-response-base/questionnaire-settings';;
-
+import type { ResponseFormQuestionnaireFormSettings } from '~/components/form-response-base/questionnaire-settings';
 const formSettings: ResponseFormQuestionnaireFormSettings = {
   title: 'アンケートのタイトル',
   description: 'アンケートの説明',
@@ -71,3 +70,4 @@ const sending = '送信';
 
 <style lang="scss" scoped></style>
 ~/components/form-response-base/questionnaire-settings
+~/components/response-form-base/questionnaire-settings
